+++ examples/customize/embeddings/vertexai_embeddings.py
@@ -4,6 +4,6 @@ using Google models and the VertexAI API.
 
 from neo4j_graphrag.embeddings import VertexAIEmbeddings
 
-embeder = VertexAIEmbeddings(model="text-embedding-004")
+embeder = VertexAIEmbeddings(model="text-embedding-005")
 res = embeder.embed_query("my question")
 print(res[:10])